feat: add shared scraper function, update Screener
- Add shared scraper function for all parsers
- Consolidate types
- Update Screener with latest additions
